requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
django-filter>=23.1
Pillow>=10.0  